Fix warnings in utils.h
In utils.h, the `isIdentical` functions for
`tesseract_scene_graph::Visual` and `tesseract_scene_graph::Collision`
generate warnings because they're non void functions with no return.
This makes including that file in a compilation unit compiled with
`-Werror` impossible.

Deleted those 2 functions as well as well as the version for
`tesseract_scene_graph::Link` as it was using them.